# contracts/openapi/gateway.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/games": {
      "get": {
        "operationId": "list_games_games_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Include PGN data in results",
            "in": "query",
            "name": "include_pgn",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include PGN data in results",
              "title": "Include Pgn",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Games",
        "tags": [
          "games"
        ]
      },
      "post": {
        "description": "Upsert a game by (provider, source_id) when provided; otherwise dedupe by digest.\nAccepts a normalized payload with fields like pgn, provider, source_id, perf, etc.",
        "operationId": "create_or_update_game_games_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Create Or Update Game",
        "tags": [
          "games"
        ]
      }
    },
...
    "/games/{game_id}/pgn": {
      "get": {
        "description": "Fetch PGN for a single game (used for lazy loading on profile/games pages).",
        "operationId": "get_game_pgn_games__game_id__pgn_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Game Pgn",
        "tags": [
          "games"
        ]
      }
    },
...
    "/import/games/fetch": {
      "post": {
        "description": "Forward import requests to import-service. Expects payload like:\n{\"source\": \"lichess.org\"|\"chess.com\", \"username\": \"...\", \"filters\": {...}}",
        "operationId": "import_fetch_games_import_games_fetch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Fetch Games",
        "tags": [
          "imports"
        ]
      }
    },
    "/import/games/fetch-by-url": {
      "post": {
        "description": "Fetch a single game by URL from Lichess or Chess.com via import-service.\nExpected payload: {\"url\": \"...\", \"source\": \"lichess.org\"|\"chess.com\"}",
        "operationId": "import_fetch_game_by_url_import_games_fetch_by_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Fetch Game By Url",
        "tags": [
          "imports"
        ]
      }
    },
    "/import/games/fetch/stream": {
      "post": {
        "description": "Stream NDJSON from import-service for incremental UI updates.",
        "operationId": "import_fetch_games_stream_import_games_fetch_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Fetch Games Stream",
        "tags": [
          "imports"
        ]
      }
    },
...
  }
}